Add comment about back, as it is helpful, but don't follow convention.
This is to make sure that the details about the output are included, but also
that the comment is not going to be parsed and used for basis of a test.

Sadly, different versions of `libmagic` often output different thing, and it
would be hard to capture this in the code comment.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>